+++ arctic_office_projects_api/main/faker/providers/project/__init__.py
@@ -1,4 +1,4 @@
-from enum import Enum
+from datetime import timedelta
 from typing import List
 
 from faker import Faker
@@ -6,12 +6,8 @@ from faker.providers import BaseProvider
 # noinspection PyPackageRequirements
 from psycopg2.extras import DateRange
 
+from arctic_office_projects_api.main.faker.providers.grant import GrantType
 
-class GrantType(Enum):
-    UKRI_STANDARD_GRANT = 'ukri-standard-grant'
-    UKRI_LARGE_GRANT = 'ukri-large-grant'
-    EU_STANDARD_GRANT = 'eu-standard-grant'
-    OTHER = 'other-grant'
 localized = False